Moves all of the functions to a separate file for cleaner code.
